Refactoring of Tests.
Tests are renamed into shorter names. Each case resides in its own file to
ensure tests run sequentially.
