+++ tests/abort2_case3.rs
@@ -2,25 +2,28 @@
 use bitcoin::Amount;
 use coinswap::{
     maker::MakerBehavior,
-    taker::TakerBehavior,
+    taker::{SwapParams, Taker, TakerBehavior},
     test_commons::*,
-    wallet::{fidelity::YearAndMonth, RPCConfig, Wallet, WalletMode},
+    wallet::{RPCConfig, WalletMode},
 };
-
 use std::{
     fs,
     path::PathBuf,
+    str::FromStr,
     sync::{Arc, RwLock},
     thread,
     time::Duration,
 };
 
-use std::str::FromStr;
-
-/// This test demonstrates a standard coinswap round between a Taker and 2 Makers. Nothing goes wrong
-/// and the coinswap completes successfully.
+/// ABORT 2: Maker Drops Before Setup
+/// This test demonstrates the situation where a Maker prematurely drops connections after doing
+/// initial protocol handshake. This should not necessarily disrupt the round, the Taker will try to find
+/// more makers in his address book and carry on as usual. The Taker will mark this Maker as "bad" and will
+/// not swap this maker again.
+///
+/// CASE 3: Maker Drops After Sending Sender's Signature. Taker and other Maker recovers.
 #[tokio::test]
-async fn test_standard_coinswap() {
+async fn maker_drops_after_sending_senders_sigs() {
     coinswap::scripts::setup_logger();
 
     let test_framework = Arc::new(TestFrameWork::new(None));
@@ -36,23 +39,25 @@ async fn test_standard_coinswap() {
     let mut maker2_rpc_config = rpc_config.clone();
     maker2_rpc_config.wallet_name = "maker_2".to_string();
 
-    // // unlock all utxos to avoid "insufficient fund" error
-    // rpc.call::<Value>("lockunspent", &[Value::Bool(true)])
-    //     .unwrap();
-
-    // create temp dir to hold wallet and .dat files if not exists
-    if !std::path::Path::new(TEMP_FILES_DIR).exists() {
-        fs::create_dir::<PathBuf>(TEMP_FILES_DIR.into()).unwrap();
+    // Start from fresh temp dir
+    if PathBuf::from_str(TEMP_FILES_DIR).unwrap().exists() {
+        fs::remove_dir_all::<PathBuf>(TEMP_FILES_DIR.into()).unwrap();
     }
 
-    // create taker wallet
-    let mut taker_wallet = create_wallet_and_import(TAKER.into(), &taker_rpc_config);
+    let mut taker = Taker::init(
+        &PathBuf::from_str(TAKER).unwrap(),
+        Some(taker_rpc_config),
+        Some(WalletMode::Testing),
+        TakerBehavior::Normal,
+    )
+    .await
+    .unwrap();
 
     // create maker1 wallet
-    let mut maker1_wallet = create_wallet_and_import(MAKER1.into(), &maker1_rpc_config);
+    let mut maker1_wallet = create_wallet_and_import(&MAKER1.into(), &maker1_rpc_config);
 
     // create maker2 wallet
-    let mut maker2_wallet = create_wallet_and_import(MAKER2.into(), &maker2_rpc_config);
+    let mut maker2_wallet = create_wallet_and_import(&MAKER2.into(), &maker2_rpc_config);
 
     // Check files are created
     assert!(std::path::Path::new(TAKER).exists());
@@ -61,7 +66,7 @@ async fn test_standard_coinswap() {
 
     // Create 3 taker and maker address and send 0.05 btc to each
     for _ in 0..3 {
-        let taker_address = taker_wallet.get_next_external_address().unwrap();
+        let taker_address = taker.get_wallet_mut().get_next_external_address().unwrap();
         let maker1_address = maker1_wallet.get_next_external_address().unwrap();
         let maker2_address = maker2_wallet.get_next_external_address().unwrap();
 
@@ -70,26 +75,16 @@ async fn test_standard_coinswap() {
         test_framework.send_to_address(&maker2_address, Amount::from_btc(0.05).unwrap());
     }
 
-    // Create a fidelity bond for each maker
-    let maker1_fbond_address = maker1_wallet
-        .get_timelocked_address(&YearAndMonth::new(2030, 1))
-        .0;
-    let maker2_fbond_address = maker2_wallet
-        .get_timelocked_address(&YearAndMonth::new(2030, 1))
-        .0;
-
-    test_framework.send_to_address(&maker1_fbond_address, Amount::from_btc(0.05).unwrap());
-    test_framework.send_to_address(&maker2_fbond_address, Amount::from_btc(0.05).unwrap());
-
     test_framework.generate_1_block();
 
     // Check inital wallet assertions
-    assert_eq!(*taker_wallet.get_external_index(), 3);
+    assert_eq!(*taker.get_wallet().get_external_index(), 3);
     assert_eq!(*maker1_wallet.get_external_index(), 3);
     assert_eq!(*maker2_wallet.get_external_index(), 3);
 
     assert_eq!(
-        taker_wallet
+        taker
+            .get_wallet()
             .list_unspent_from_wallet(false, true)
             .unwrap()
             .len(),
@@ -100,17 +95,20 @@ async fn test_standard_coinswap() {
             .list_unspent_from_wallet(false, true)
             .unwrap()
             .len(),
-        4
+        3
     );
     assert_eq!(
         maker2_wallet
             .list_unspent_from_wallet(false, true)
             .unwrap()
             .len(),
-        4
+        3
     );
 
-    assert_eq!(taker_wallet.lock_all_nonwallet_unspents().unwrap(), ());
+    assert_eq!(
+        taker.get_wallet().lock_all_nonwallet_unspents().unwrap(),
+        ()
+    );
     assert_eq!(maker1_wallet.lock_all_nonwallet_unspents().unwrap(), ());
     assert_eq!(maker2_wallet.lock_all_nonwallet_unspents().unwrap(), ());
 
@@ -124,7 +122,7 @@ async fn test_standard_coinswap() {
             &maker1_config_clone,
             6102,
             Some(WalletMode::Testing),
-            MakerBehavior::Normal,
+            MakerBehavior::CloseAfterSendingSendersSigs,
             kill_flag_maker_1,
         )
         .unwrap();
@@ -144,87 +142,39 @@ async fn test_standard_coinswap() {
         .unwrap();
     });
 
-    let taker_config_clone = taker_rpc_config.clone();
-    let taker_thread = thread::spawn(|| {
-        // Wait and then start the taker
-        thread::sleep(Duration::from_secs(20));
-        coinswap::scripts::taker::run_taker(
-            &PathBuf::from_str(TAKER).unwrap(),
-            Some(WalletMode::Testing),
-            Some(taker_config_clone), /* Default RPC */
-            1000,
-            500000,
-            2,
-            3,
-            TakerBehavior::Normal,
-        );
-    });
-
     let test_frameowrk_ptr = test_framework.clone();
     let kill_block_creation_clone = kill_flag.clone();
     let block_creation_thread = thread::spawn(move || {
         while !*kill_block_creation_clone.read().unwrap() {
-            thread::sleep(Duration::from_secs(5));
+            thread::sleep(Duration::from_secs(1));
             test_frameowrk_ptr.generate_1_block();
             log::info!("created block");
         }
         log::info!("ending block creation thread");
     });
 
-    taker_thread.join().unwrap();
+    let swap_params = SwapParams {
+        send_amount: 500000,
+        maker_count: 2,
+        tx_count: 3,
+        required_confirms: 1,
+        fee_rate: 1000,
+    };
+
+    thread::sleep(Duration::from_secs(20));
+    taker.send_coinswap(swap_params).await.unwrap();
     *kill_flag.write().unwrap() = true;
     maker1_thread.join().unwrap();
     maker2_thread.join().unwrap();
+
     block_creation_thread.join().unwrap();
 
-    // Recreate the wallet
-    let taker_wallet =
-        Wallet::load(&taker_rpc_config, &TAKER.into(), Some(WalletMode::Testing)).unwrap();
-    let maker1_wallet = Wallet::load(
-        &maker1_rpc_config,
-        &MAKER1.into(),
-        Some(WalletMode::Testing),
-    )
-    .unwrap();
-    let maker2_wallet = Wallet::load(
-        &maker2_rpc_config,
-        &MAKER2.into(),
-        Some(WalletMode::Testing),
-    )
-    .unwrap();
+    // Maker gets banned for being naughty.
+    assert_eq!(
+        "localhost:6102",
+        taker.get_bad_makers()[0].address.to_string()
+    );
 
-    // Check assertions
-    assert_eq!(taker_wallet.get_swapcoins_count(), 6);
-    assert_eq!(maker1_wallet.get_swapcoins_count(), 6);
-    assert_eq!(maker2_wallet.get_swapcoins_count(), 6);
-
-    let utxos = taker_wallet.list_unspent_from_wallet(false, false).unwrap();
-    let balance: Amount = utxos
-        .iter()
-        .fold(Amount::ZERO, |acc, (u, _)| acc + u.amount);
-    assert_eq!(utxos.len(), 6);
-    assert!(balance < Amount::from_btc(0.15).unwrap());
-
-    let utxos = maker1_wallet
-        .list_unspent_from_wallet(false, false)
-        .unwrap();
-    let balance: Amount = utxos
-        .iter()
-        .fold(Amount::ZERO, |acc, (u, _)| acc + u.amount);
-    assert_eq!(utxos.len(), 6);
-    assert!(balance > Amount::from_btc(0.15).unwrap());
-
-    let utxos = maker2_wallet
-        .list_unspent_from_wallet(false, false)
-        .unwrap();
-    let balance: Amount = utxos
-        .iter()
-        .fold(Amount::ZERO, |acc, (u, _)| acc + u.amount);
-    assert_eq!(utxos.len(), 6);
-    assert!(balance > Amount::from_btc(0.15).unwrap());
-
-    // Remove test temp files and dir
     fs::remove_dir_all::<PathBuf>(TEMP_FILES_DIR.into()).unwrap();
-
     test_framework.stop();
 }
